Refactor: Organize repositories into separate packages
