+++ src/main/java/com/activecourses/upwork/repository/auth/RefreshTokenRepository.java
@@ -1,4 +1,4 @@
-package com.activecourses.upwork.repository;
+package com.activecourses.upwork.repository.auth;
 
 import com.activecourses.upwork.model.RefreshToken;
 import com.activecourses.upwork.model.User;
