+++ src/main/java/com/activecourses/upwork/service/authentication/RefreshTokenService.java
@@ -4,8 +4,8 @@ import com.activecourses.upwork.config.security.jwt.JwtService;
 import com.activecourses.upwork.dto.ResponseDto;
 import com.activecourses.upwork.exception.TokenRefreshException;
 import com.activecourses.upwork.model.RefreshToken;
-import com.activecourses.upwork.repository.RefreshTokenRepository;
-import com.activecourses.upwork.repository.UserRepository;
+import com.activecourses.upwork.repository.auth.RefreshTokenRepository;
+import com.activecourses.upwork.repository.user.UserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
